refactor: shift `TaskType` under evaluator model
