+++ unicon_backend/evaluator/tasks/base.py
@@ -5,11 +5,17 @@ from typing import Any, Generic, TypeVar
 from pydantic import BaseModel
 
 from unicon_backend.lib.common import CustomBaseModel
-from unicon_backend.models.contest import TaskType
 
 TaskUserInput = TypeVar("TaskUserInput")
 TaskExpectedAnswer = TypeVar("TaskExpectedAnswer")
 TaskResult = TypeVar("TaskResult")
+
+
+class TaskType(str, Enum):
+    MULTIPLE_CHOICE = "MULTIPLE_CHOICE_TASK"
+    MULTIPLE_RESPONSE = "MULTIPLE_RESPONSE_TASK"
+    SHORT_ANSWER = "SHORT_ANSWER_TASK"
+    PROGRAMMING = "PROGRAMMING_TASK"
 
 
 class TaskEvalStatus(str, Enum):
